requirements.txt
Flask==2.3.2
psycopg2-binary==2.9.5
python-dotenv==1.0.0
google-cloud-spanner==3.26.0